Refactor import paths: Update Loader component imports for consistency across files
