+++ app/src/components/learn/lesson/Lesson.jsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { useEffect, useState } from "react";
-import Loader from "../../../components/loader/Loader";
+import Loader from "../../loader/Loader";
 import API from "../../../utils/api";
 import { motion, AnimatePresence } from "framer-motion";
 import { FaStar } from "react-icons/fa";
